xatmi - comply to specification signal handling
Before: We did not correctly handle signals according to the
specification for the TPSIGRSTRT flag.
Nor did we set tperrno to TPGOTSIG if a call failed due to
receiving a signal from the OS.

Now: For the XATMI functions where TPSIGRSTRT is valid, we now
block incoming signals for the duration of the call. If TPSIGRSTRT
is not set and a call fails due to a signal, we now set tperrno to
TPGOTSIG.

Resolves #271