http: use a single correlation for entire service call
Before: the http-inbound used different correlation ids for looking
up and calling services, contrary to what the service-manager expected.
This caused deadlines for services with timeouts not to be removed,
which then expired even though the call had previously completed.
Now: the http-inbound uses the same correlation id for the entire
service call.

Resolves #318