service: pid -> ipc  & route calls to same instance based on gtrid
We now keep track of instances via their ipc-address instead of pid.
We also keep track of gtrid of each lookup request and to which
instance/ipc we reply with. This enables SM to let as much calls as
possible go to the same remote instance based on gtrid. That is, keep
the transaction as little distributed as possible.

issue #238