gateway: fix markdown documentation for interdomain protocol
Before: all int (32b scalars) where generated as uint64 in the
documentation. Although, the actual bits that are sent over the wire
are uint32.

I'm not sure if sending 32b where intentional. But this is what we've
got since 1.4, and we can't (of course) change the protocol.

Now: The markdown documentation (generator) is producing the correct
documentation for 32b types.

resolves #319