common: do not set environment when (re)setting execution-id
Before: We set an environment variable when (re)setting exection-id to
propagate this to potential children. Setting environment variables
leads to memory leaks on most linux (which in it self is very odd).

Now: We add an environment variable when we spawn a child, in the spawn
api. Hence, we don't alter parent variable state -> no memory leaks

Fixes #333