common: improve flag abstraction to use only pure enums
A lot cleaner and quite "clever" (subjective) reduction of the
abstraction.

Using _declare a function with a specific name_ to opt in if a `enum`
should be treated as a flag-like. Quite terse, and one can opt in
directly where the `enum` is defined. Using "classic"
type-specialization to opt in, one (often) need to opt in "far" from
the type. I'm not sure which is better... time will tell.

You can also opt in and use another flag as superset:
`consteval <the superset type> casual_enum_as_flag_superset( <your enum type>);

This is helpful since we've split relevant flag values into distinct
flag types for a given function. That is, for a given function a subset
of the possible flags are valid. We define this subset as it's own enum
type. `casual_enum_as_flag_superset` opt in will then use the superset
enum `description` function.