unittest: cli::command does not pass correct environment variables
Before: We used `popen` under the hood which of course only passed
system environment variables.
Now when we have another set of variables that we pass on via
`posix_spawn` with our regular process::spawn (used every where else),
we need to pass them to the cli command, to make it work.

Now: cli::command uses the improved process::execute that passes our
own mutable environment, just as process::spawn. The command get the
correct variables in its environment.

`cli::command` returns a type `Capture` that has the `exit code` of the
process, and standard out and error strings. This makes it a lot easier
to test cli stuff. Now we can actually test that something was printed
to stderr.

Relates #335