discovery: short-circuit fan-out discoveries -> reply directly on found
Before: When a discovery was made to all providers (outbound to other
domains) we did a fan-out and waited for all replies until we replied to
the inbound request (another domain, SM, QM asking discovery).
If the platform has problem with the network stack, and other system
related problems, we could wait for ever. And new discoveries would also
be stuck in the waiting game. This could have negative effects on a
deployed casual topology.

Now: We short-circuit the fan-out if we've discovered the requested
services/queues somewhere, and reply that we know of it.
This should:
* Make casual more robust (arcane network stack problem)
* Reply to discovery requests faster

One could argue that arcane network stack problems that casual
could not detect is a thing that is out-of-scope to mitigate for
`casual`. But it feel semantically correct to reply to a discovery
request as soon as we know that the wanted services/queues exists in one
of the registered providers.

Resolves #372