common: archive type -> properties (flag) order/named at the same time
Before: Archives implemented archive_type() that could only be _named_
or _order_, not both.

Now: Archives implements archive_properties() that return an _enum
flag_, that can contain _named_ and _order_.

This is mostly to be able to create correct markdown documentation for
interdomain protocol.

With this we know we get same serialization mechanism (the same values)
during documentation generation as we get when we do the actual binary
serialization that goes over the wire.