gateway: new enqueue/dequeue-reply messages version (code)
This patch adds new versions for enqueue::Reply and dequeue::Reply.
* code/error to be able to propagate errors
* message is an optional instead of std::vector

Also, the mechanism to be able to limit message compatibility to a
min/max versions. That is, to be able to say that a message is used up
to a given version.

Also, exposed _protocol version_ to the admin model and CLI.