refactoring: binary type std::vector< char> -> std::vector< std::byte>
This is a major refactoring, since we use platform::binary::type in a
lot of places.

Why? To have a distinct type that represent binary is clearer and less
error prone. std::as_bytes can be used seamless to handle trivially
copyable types to and from binary.
Also the binary type does not look like a string-like type, and should
not be used as such (which we did/do) in some places.

A lot of hacking in `buffer` (fielded & CO) where the implementation
is very C:ish. I didn't want to rewrite the whole thing...