queue: local domain enable/disable enqueue/dequeue
This patch implements a local domain enable/disable for enqueue/dequeue
per queue.

This enable users to disable enqueue/dequeue for a given queue. The
semantics for _disabling_ are equal to 'hide' the queue for a given
action (enqueue/dequeue). That is, if a lookup occur with the action
'enqueue' for a queue that has _enqueue_ disabled, the lookup will reply
with _no-queue_, as if the queue does not exists.

This is the first part of this. We need to extend the discovery 'reply'
to include the _enable_ of the queue. Hence, have the same semantics
between domain. I think we'll cram this in interdomain protocol 1.3.