service: removed intermediate service lookup reply
Before: If a requested service was busy, service-manager replied with an
intermediate busy reply, that hold the same information as the idle
reply later on. This was to enable caller to set "timers" and whatnot.
This has not been used for a while, so this patch got rid of it.