common: execution id should be inherited by children
Before: When execution id was reset, we did not update/set the
corresponding (internal) environment variable -> every child process
from a unittest (domains and so on) had it's own execution id. This
makes it unnecessary hard to filter out a particular unittest from a log
file with hundreds of unittests.

Now: We set the (internal) environment virable.

Note: We don't actually set to the system environment variables, we set
to our own _prioritized_ set of variables, that child processes gets
during spawn.