refactoring: service timeout -> service call deadline
The naming _call deadline_ makes it clearer that the deadline is not
necessary bound to whatever timeout the service has.
A service call gets the earliest deadline (timepoint) of the deadline it
already has (if any) and the timeout of the requested service.

Also: Cleaned up common/message/service.h. There's different needs to
what a _service_ represent in different context.
Before: we had a lot of inheritance (code reuse) which made it
unnecessary hard to maintain the different needs over time. Unnecessary
attributes got sucked in, which made it harder to understand, besides
the fact that we sent unnecessary bytes.
Now: Every _service type_ is it's own bespoke type.

issue #296