service: refactoring service-manager - reference based -> index based
Before: services and instances had their circular dependencies made of
pointers and references. This made it a bit error prune and hard to
reason about.

Now: Every service has an id, every instance has an id. The relation
between them is just "a vector with ids". This is:
* a service has all ids of instances that has advertise the service
* an instance has all the ids of services that the instance has
  advertised.

Also, added a _disabled_ set that holds all instances that is in
_prepare for shutdown_ mode. This enables us to keep track of _disabled_
instances and be able to handle a timeout properly during this mode.

Resolves #407