cli: improved --legend <option> to --legend --<option>
Before: --legend had to be invoked with a different name than the actual
option name. This to prevent the parsing to interpret the passed value
as an actual invocation of the value as the option.
`casual service --legend list-services`

Now: We use suboptions for --legend with the name of the actual option
`casual service --legend --list-services`

As of now, we can't force at least one suboption to be passed -> user
could get completion for other actual options that does not have
legends. This could happen _before_ also.

It would be nice to have `at_least_one( subopt1, subopt2, ..., suboptN),
but I'm not sure how that would work in all situations.