discovery: handle discoverable::Advertised message -> topology::Update
This patch adds the message discoverable::Advertised. When discovery
receives this message it gets added to _topology accumulation_. When the
timer kicks in, a `topology::Update` is sent to _inbounds_ for further
propagation upstream.

This enables a domain to notify upstream domains that there could be new
services/queues to be discovered.

resolves #479