http: user localhost instead of a.example as host in unittests
Before: we used a.example to indicate an unreachable host/url. This, for
some reason, could render a long timeout (several minutes). It's not
clear what exactly is the root of the problem. It could be some curl
stuff, or it could be some OS dns cache stuff.

Now: we use `localhost` as host, whcih is known to the OS. This should
fix the long timeouts. However, this might call a webserver that is
listening on localhost, but the url should be unique enough to still at
least get a 404, and unittests get the expected result.