http: exception wrapper -> return code within C-boundary for xatmi call
Before: if an exception is thrown the nginx worker process will core.
This could happen if no content-type header was provided, for example.

Now: We have an exception guard that returns the _error-code_. For now
this result in an NGX_ERROR.

We might want to give a more precise http-error, based on the exception.

Resolves #487