gateway: outbound does not set otel span in metric event
Before: otbound did not create an otel span for service calls and did
not provide span in the metric event.

Now: we generate an otel span for each call, and use the call message
parent span, as parent span in metric event.
We set outbound otel span as the new parent span in the call message for
downstream.

Resolves #539