cleanup: reduced the number of log-categories
Before: We had a lot of module-specific log categories that did not help
us that much. "all" of it was debug logging anyway.

Now: We use log::debug, and user has to enable debug logging on the
"manager" of interest.

Going forward, I think we only use _debug_ for debug stuff, and
_event-logging_ for stuff that we think is important enough.