common: server-service finalize needs to wait for pending calls
Before: The default Policy (that is used for user servers) didn't
wait/fetch pending in-flight replies before sending rollback to TM. This
could lead to timing issues if TM reaches resources before any of the
pending calls.

Now: We wait for all in-flight calls that are associated with a
transaction.

Resolves #575