server: cleaned up server::context handlers
This patch:
* server::context::initialize returns handlers for service call and
  conversation connect. Used when initialize 'start'
* Got rid of all _policy_ stuff -> we don't need any polices since
  managers don't use this any more.

There are a lot of clean up that still could be done. Especially the
state that cobol needs. It´should be something that "only cobol" knows
about.
With this patch I think it's going to be a bit easier to improve the
code going forward. It's more isolated and so on.