manager: added concurrent service invocation functionality
This patch adds functionality to construct concurrent services. The
manager service implementation takes a callback object that it calls
when the service is done. This callback will send ack (depends on what
the policy does) and send reply to caller (if the caller wants a reply).

For now, these concurrent services are still advertised as sequential,
since it's not clear what effect multiple concurrent calls would have.
It should work fine, but I'm not totally sure.
Also, there's the situation if a service is not found on the server,
what _ack_ should we send to SM? We don't know if the service is
concurrent or sequential. We may need to provide the
sequential/concurrent in the server-call-request.

There are still benefits even when the concurrent services are
advertised as sequential.
* caller can block until the call-reply arrives, and know that the
  service is done.
* We got rid of all the waiting for events to "know" when the logical
  service is done.

Resolves #584