domain: if server exit during boot the task might get stuck
Before: When a server instance exit during boot/startup and the server
has `restart`. The state for the instance remains _scale-out_, but no
restart will happen (due to boot/startup runlevel). This will render the
task 'stuck', and a lot of things will not work (all stuff that also
uses task to get their job done, since they're blocked by the stuck
task.

Now: If we're in another runleveln than `running` we just set wanted
'state' of the instance to _linger_, which the task will consider to be
done. The instance state becomes `error` in the cli, which is the
correct state for the instance in this scenario.

Resolves #624