service: SM doesn't remove pending deadlines when an instance dies
Before: SM did not remove associated pending deadlines to an exited
instance -> when the deadline kicks in SM fails to correlate the
deadline and error log.

Now: SM removes associated pending deadlines when an instance exit

Had to expand admin-state-model to include _pending-deadlines_, to be
able to test it. This implies that we've not tested this before. Better
late than never...

Resolves #630