tools: call cli don't use the same correlation for call as lookup
Before: the cli for `casual call ...` didn't use the same correlation
for the call message as provided for the service lookup message.
If a scale-in/shutdown of the callee is done mid-flight and callee
sends ACK to SM, SM can't correlate the ACK and no
shutdown-request-reply will be sent to DM, and the scale-in/shutdown
gets "stuck"

Now: we make sure to use the same correlation-id for both service-lookup
and service-call messages.