service: possible UB in remove deadline - deadline improvements
This patch: When we removed element in pending.deadline we could dereference
`deadline.end()` which is `UB`. But the memory is still there so the
consequence is "defined". Still, it's `UB` and this patch fixes it.

Added semantics to unset the current timer (alarm) if the removed
pending.deadline was the last. This eliminates that we get an alarm
signal but no deadlines. Nothing happens when the stale alarm signal is
processed, but it's still unnecessary.