http: inbound does not discard service lookup on some error paths
Before: When a new inbound::call::Context was instantiated, the service
lookup was sent regardless. If the validation of the request failed
(missing mandatory headers and such) the service lookup was never
discarded -> the service became reserved "for ever" (depending on
timeouts and such)

Now: We make sure to always send lookup-discard if the service-lookup is
in flight, regardless of errors.

Resolves: #647